scripts/simulated-sensors/motion_sensor/requirements.txt
paho-mqtt==1.6.1
kafka-python
psycopg2-binary